+++ e2e/test/babel-transpiling/stryker.conf.js
@@ -6,7 +6,7 @@ module.exports = {
   coverageAnalysis: 'off',
   buildCommand: 'npm run build',
   timeoutMS: 60000,
-  reporters: ['clear-text', 'html', 'event-recorder', 'progress'],
+  reporters: ['clear-text', 'html', 'json', 'progress'],
   concurrency: 1,
   logLevel: 'info',
   plugins: [