+++ e2e/test/cucumber-ts/cucumber.js
@@ -1,3 +1,3 @@
 module.exports = {
-  default: '--require test/cucumber-ts/features/support/setup.js --require features/**/*.ts --publish-quiet'
+  default: '--require features/support/setup.js --require features/**/*.ts --publish-quiet'
 }