+++ packages/core/src/child-proxy/child-process-proxy-worker.ts
@@ -26,8 +26,8 @@ export class ChildProcessProxyWorker {
       process.send(str);
     }
   }
-  private handleMessage(serializedMessage: string) {
-    const message = deserialize<WorkerMessage>(serializedMessage);
+  private handleMessage(serializedMessage: unknown) {
+    const message = deserialize<WorkerMessage>(String(serializedMessage));
     switch (message.kind) {
       case WorkerMessageKind.Init:
         this.handleInit(message);
@@ -51,7 +51,7 @@ export class ChildProcessProxyWorker {
     this.log = getLogger(ChildProcessProxyWorker.name);
     this.handlePromiseRejections();
     let injector = buildChildProcessInjector(message.options);
-    const locals = message.additionalInjectableValues as Record<string, unknown>;
+    const locals = message.additionalInjectableValues as Record<string, any>;
     for (const token in locals) {
       injector = injector.provideValue(token, locals[token]);
     }