+++ packages/core/src/initializer/stryker-inquirer.ts
@@ -43,6 +43,18 @@ export class StrykerInquirer {
     }
   }
 
+  public async promptBuildCommand(skip: boolean): Promise<PromptOption> {
+    const { buildCommand } = await inquirer.prompt<{ buildCommand: string }>({
+      message:
+        'What build command should be executed just before running your tests? For example: "npm run build" or "tsc -b" (leave empty when this is not needed).',
+      name: 'buildCommand',
+      default: 'none',
+      when: !skip,
+    });
+
+    return { name: buildCommand !== 'none' ? buildCommand : '', pkg: null };
+  }
+
   public async promptReporters(options: PromptOption[]): Promise<PromptOption[]> {
     const answers = await inquirer.prompt<{ reporters: string[] }>({
       choices: options.map((_) => _.name),