+++ packages/core/src/stryker-cli.ts
@@ -28,7 +28,7 @@ function deepOption<T extends string, R>(object: { [K in T]?: R }, key: T) {
 const list = createSplitter(',');
 
 function createSplitter(sep: string) {
-  return (val: string) => val.split(sep);
+  return (val: string) => val.split(sep).filter(Boolean);
 }
 
 function parseBoolean(val: string) {
@@ -85,6 +85,12 @@ export class StrykerCli {
         'Configure a build command to run after mutating the code, but before mutants are tested. This is generally used to transpile your code before testing.' +
           " Only configure this if your test runner doesn't take care of this already and you're not using just-in-time transpiler like `babel/register` or `ts-node`."
       )
+      .option(
+        '--checkers <listOfCheckersOrEmptyString>',
+        'A comma separated list of checkers to use, for example --checkers typescript',
+        createSplitter(',')
+      )
+      .option('--checkerNodeArgs <listOfNodeArgs>', 'A list of node args to be passed to checker child processes.', createSplitter(' '))
       .option(
         `--coverageAnalysis <perTest|all|off>', 'The coverage analysis strategy you want to use. Default value: "${defaultValues.coverageAnalysis}"`
       )
@@ -163,6 +169,7 @@ export class StrykerCli {
         `Choose whether or not to clean the temp dir (which is "${defaultValues.tempDirName}" inside the current working directory by default) after a successful run. The temp dir will never be removed when the run failed for some reason (for debugging purposes).`,
         parseBoolean
       )
+      .showSuggestionAfterError()
       .parse(this.argv);
 
     // Earliest opportunity to configure the log level based on the logLevel argument