+++ packages/jasmine-runner/typings/global.d.ts
@@ -1,7 +1,4 @@
-declare namespace NodeJS {
-  interface Global {
-    __stryker__: import('@stryker-mutator/api/core').InstrumenterContext | undefined;
-    // used during testing, so we can actually run stryker on the @stryker-mutator/jasmine-runner package itself
-    __stryker2__: import('@stryker-mutator/api/core').InstrumenterContext | undefined;
-  }
-}
+/* eslint-disable no-var */
+// Use "var" instead of let, otherwise we are not expanding the `globalThis`
+declare var __stryker__: import('@stryker-mutator/api/core').InstrumenterContext | undefined;
+declare var __stryker2__: import('@stryker-mutator/api/core').InstrumenterContext | undefined;