+++ packages/mocha-runner/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.5.1](https://github.com/stryker-mutator/stryker-js/compare/v5.5.0...v5.5.1) (2021-12-02)
+
+**Note:** Version bump only for package @stryker-mutator/mocha-runner
+
+
+
+
+
+# [5.5.0](https://github.com/stryker-mutator/stryker-js/compare/v5.4.1...v5.5.0) (2021-11-23)
+
+
+### Features
+
+* **hit limit:** infinite loop prevention in mocha-runner  ([f5a7d1d](https://github.com/stryker-mutator/stryker-js/commit/f5a7d1d18ec45364743e5aceb71f0f1bbbf3bafa))
+
+
+
+
+
 ## [5.4.1](https://github.com/stryker-mutator/stryker-js/compare/v5.4.0...v5.4.1) (2021-09-30)
 
 **Note:** Version bump only for package @stryker-mutator/mocha-runner